+++ modules/showLikes.js
@@ -1,0 +1,8 @@
+import fetchAPI from './fetchAPI';
+
+const showLikes = async () => {
+  const likes = await fetchAPI('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/PVsyyzJcjAI3kwRWNz7v/likes');
+  console.log(likes);
+};
+
+export default showLikes;
