+++ src/styles.css
@@ -158,6 +158,7 @@ i.fa-regular.fa-heart {
   top: 10px;
   right: -40px;
   transform: rotate(45deg);
+  cursor: pointer;
 }
 
 .vertical-line,
